scan(cuda): Add Blelloch monoblock scan kernel
- Implement single-block Blelloch (up-sweep, down-sweep) with support
for both exclusive and inclusive scan via shared-memory shift +
register-cached total. Requires power-of-2 input size.

- Route Blelloch through scan_cuda_device alongside Hillis-Steele,
using half the shared memory (no double buffer).

- Add tests validating Blelloch inc/exc against Hillis-Steele and serial results.

+++ 03_scan/include/mx_scan/detail/device/hillis_steele.cuh
@@ -5,10 +5,11 @@
 
 namespace mx::detail {
 
+// Only works for pow2 problem sizes that fit within a thread block i.e. <= 1024
 template<ScanType scan_type, typename T, typename Op>
 __global__ void hillis_steele_on_device_monoblock(const T* __restrict__ input,
-                                                                 T* __restrict__ output,
-                                                                 size_t size, Op op)
+                                                  T* __restrict__ output,
+                                                  int size, Op op)
 {
     int tid = threadIdx.x; // 1 block, 1D grid
